requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
pyyaml>=6.0.1
schedule>=1.2.1
fake-useragent>=1.4.0